requirements.txt
spotipy>=2.23.0
flask>=3.0.0
discord.py>=2.3.0
python-dotenv>=1.0.0
yt-dlp>=2023.12.30
aiohttp>=3.9.0
